Remove duplicated repositories
In case of bare or standard repositories engine can return more than 1
repository for the same directory.

The most common case is when you just "git clone" remote repository and
run gemini on it.

Engine would produce 2 different repositories:
- local one: file://path/to/directory
- remote one: https://github.com/...

As a result all files in the repository will have duplicates which is
incorrect from user point of view.

Signed-off-by: Maxim Sukharev <max@smacker.ru>